% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{call_llm_api}
\alias{call_llm_api}
\title{Call the LLM API with the prompt}
\usage{
call_llm_api(prompt)
}
\arguments{
\item{prompt}{Prompt to use as input}
}
\description{
Call the LLM API with the prompt
}
