% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{get_platform_specific_advice}
\alias{get_platform_specific_advice}
\title{Retrieve post-writing advice unique to specific platforms}
\usage{
get_platform_specific_advice(platforms)
}
\arguments{
\item{platforms}{Which platforms to get advice for}
}
\description{
Retrieve post-writing advice unique to specific platforms
}
