% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{get_prompt}
\alias{get_prompt}
\title{Construct LLM prompt}
\usage{
get_prompt(blog_link, platforms, n, emojis, tone, hashtags)
}
\arguments{
\item{blog_link}{URL of source material}

\item{platforms}{Social media platform to create prompts for}

\item{n}{Number of prompts to create for each platform}

\item{emojis}{Use emojis in post?}

\item{tone}{Desired tone of the post}

\item{hashtags}{Hashtags to include in the post}
}
\description{
Construct a LLM prompt based on user input
}
